meallo-be/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
environs==14.1.1
fastapi==0.115.12
fastapi_cors==0.0.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
marshmallow==3.26.1
motor==3.7.0
multidict==6.2.0
packaging==24.2
propcache==0.3.1
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
twilio==9.5.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
